feat: refactor routing structure, implement public and protected routes, and add first page logic
